+++ location/admin.py
@@ -1,3 +1,11 @@
 from django.contrib import admin
 
 # Register your models here.
+from location.models import Restaurant
+
+
+class RestaurantAdmin(admin.ModelAdmin):
+    model = Restaurant
+
+
+admin.site.register(Restaurant, RestaurantAdmin)
