+++ location/admin_api/serializers.py
@@ -1,5 +1,6 @@
 from rest_framework import serializers
 from location.models import Location, Restaurant
+from location.admin_api import create_restaurant
 
 
 class LocationModelSerializer(serializers.ModelSerializer):
@@ -12,3 +13,7 @@ class RestaurantModelSerializer(serializers.ModelSerializer):
     class Meta:
         model = Restaurant
         fields = ["name", "owner", "address", "available_in", "code", "in_school"]
+        extra_kwargs = {"code": {"read_only": True}}
+
+    def save(self):
+        create_restaurant(**self.validated_data)
