+++ location/admin_api/services.py
@@ -10,6 +10,8 @@ def create_restaurant(
     address: str,
     available_in: object,
     in_school: bool,
+    *args,
+    **kwargs,
 ):
     restaurant = Restaurant(
         name=name,
