+++ utils/mixins.py
@@ -1,5 +1,9 @@
 from django.db import models
 
+from rest_framework import viewsets
+from rest_framework import status
+from rest_framework.response import Response
+
 
 class TimeStampMixin(models.Model):
     "determins when model was created and edited"
@@ -16,3 +20,21 @@ class TimeStampMixin(models.Model):
 
     def __str__(self):
         return self.name
+
+
+class ModelMixins(viewsets.ModelViewSet):
+    def create(self, request, *args, **kwargs):
+        serializer = self.get_serializer(data=request.data)
+        if serializer.is_valid():
+            self.perform_create(serializer)
+            headers = self.get_success_headers(serializer.data)
+            return Response(
+                {"success": True, "message": "successful", "data": serializer.data},
+                status=status.HTTP_201_CREATED,
+                headers=headers,
+            )
+        else:
+            return Response(
+                {"success": False, "message": "failed", "data": serializer.errors},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
